+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Version 0.6.4
 
+- Added analysis.nTE() function to calculate normalized transfer entropy (issue #42)
+
 - Added option to fix soma depth for subcellular distributions based on 1d and 2d maps
 
 - Improved import of multicompartmental cells from NeuroML2